Ajustando retorno das requisições

+++ src/main/java/com/monitoramento/saude/service/AlimentoService.java
@@ -6,7 +6,10 @@ import com.monitoramento.saude.dto.AlimentoResponseDTO;
 import com.monitoramento.saude.model.Alimento;
 import com.monitoramento.saude.repository.AlimentoRepository;
 import jakarta.persistence.EntityNotFoundException;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
+
+import java.util.Optional;
 
 @Service
 public class AlimentoService {
@@ -27,8 +30,15 @@ public class AlimentoService {
     }
 
 
-    public void deleteAlimento(Long id) {
-        repository.deleteById(id);
+    public ResponseEntity<String> deleteAlimento(Long id) {
+        Optional<Alimento> alimento = repository.findById(id);
+
+        if (alimento.isEmpty()) {
+            throw new EntityNotFoundException("Alimento não encontrado");
+        } else {
+            repository.deleteById(id);
+            return ResponseEntity.ok("Alimento com o ID " + id + " excluído com sucesso!");
+        }
     }
 
     public AlimentoResponseDTO editarAlimento(Long id, AlimentoRequestDTO dados) {
